requirements.txt
cairocffi==1.6.1
CairoSVG==2.7.1
cffi==1.15.1
cssselect2==0.7.0
defusedxml==0.7.1
deprecation==2.1.0
lxml==4.9.3
packaging==23.1
pdf2image==1.16.3
pikepdf==8.4.0
Pillow==10.0.0
pycparser==2.21
tinycss2==1.2.1
webencodings==0.5.1
